add auth and permissions

+++ app/Http/Controllers/ClientController.php
@@ -21,8 +21,17 @@ class ClientController extends Controller
      */
     public function index()
     {
-        $clients = $this->clientService->getClients();
-        return view("clients.index", ["clients" => $clients]);
+        if(\Auth::check()){
+            $authUser = \Auth::user();
+            if($authUser->can("viewAny", Client::class)){
+                $clients = $this->clientService->getClients();
+                return view("clients.index", ["clients" => $clients]);
+            } else {
+                return "Can't";
+            }
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -32,7 +41,16 @@ class ClientController extends Controller
      */
     public function create()
     {
-        return view("clients.create");
+        if(\Auth::check()){
+            $authUser = \Auth::user();
+            if($authUser->can("create", Client::class)){
+                return view("clients.create");
+            } else {
+                return "Can't";
+            }
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -43,9 +61,18 @@ class ClientController extends Controller
      */
     public function store(Request $request)
     {
-        $data = $request->only(["name", "last_name", "patronymic", "interest_status", "email", "addres", "phone", "wishes", "complaints", "selected_service", "user_id"]);
-        $this->clientService->createClient($data);
-        return redirect()->back();
+        if(\Auth::check()){
+            $authUser = \Auth::user();
+            if($authUser->can("create", Client::class)){
+                $data = $request->only(["name", "last_name", "patronymic", "interest_status", "email", "addres", "phone", "wishes", "complaints", "selected_service", "user_id"]);
+                $this->clientService->createClient($data);
+                return redirect()->back(); 
+            } else {
+                return "Can't";
+            }
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -56,8 +83,17 @@ class ClientController extends Controller
      */
     public function show($id)
     {
-        $client = $this->clientService->getClient($id);
-        return view("clients.show", ["client" => $client]);
+        if(\Auth::check()){
+            $authUser = \Auth::user();
+            $client = $this->clientService->getClient($id);
+            if($authUser->can("view", $client)){
+                return view("clients.show", ["client" => $client]);
+            } else {
+                return "Can't";
+            }
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -68,8 +104,17 @@ class ClientController extends Controller
      */
     public function edit($id)
     {
-        $client = $this->clientService->getClient($id); 
-        return view("clients.edit", ['client' => $client]);
+        if(\Auth::check()){
+            $authUser = \Auth::user();
+            $client = $this->clientService->getClient($id); 
+            if($authUser->can("update", $client)){
+                return view("clients.edit", ['client' => $client]);
+            } else {
+                return "Can't";
+            }
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 
     /**
@@ -81,9 +126,17 @@ class ClientController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $data = $request->only(["name", "last_name", "patronymic", "interest_status", "email", "addres", "phone", "wishes", "complaints", "selected_service", "user_id"]);
-        $this->clientService->updateClient($data, $id);
-        return redirect()->back();
+        if(\Auth::check()){
+            $data = $request->only(["name", "last_name", "patronymic", "interest_status", "email", "addres", "phone", "wishes", "complaints", "selected_service", "user_id"]);
+            $authUser = \Auth::user();
+            $updatingClient = $this->clientService->getClient($id);
+            if($authUser->can("update", $updatingClient)){
+                $this->clientService->updateClient($data, $id);
+                return redirect()->back();
+            } else {
+                return "Can't";
+            }
+        }
     }
 
     /**
@@ -94,7 +147,17 @@ class ClientController extends Controller
      */
     public function destroy($id)
     {
-        $this->clientService->deleteClient($id);
-        return redirect()->back();
+        if(Auth::check()){
+            $authUser = \Auth::user();
+            $updatingClient = $this->clientService->getClient($id);
+            if($authUser->can("delete", $updatingKnowledge)){
+                $this->clientService->deleteClient($id);
+                return redirect()->back();
+            } else {
+                return "Can't";
+            }
+        } else {
+            return redirect()->route('user.login');
+        }
     }
 }
